Completed first draft of Fluent API
Completed the implementation and testing of first draft of fluent API
