+++ EventSourceProxy/Fluent/Anything.cs
@@ -1,12 +1,11 @@
-﻿using System;
-
+﻿
 namespace EventSourceProxy.Fluent
 {
     public static class Anything
     {
         public static ITraceDescriptionForParam<TParam> With<TParam>()
         {
-            throw new NotImplementedException();
+            return new TraceDescriptionForParam<TParam>();
         }
     }
 }
